requirements.txt
Flask
firebase-admin
pymongo
requests
python-dotenv